fixSaveMethod

+++ model/src/main/java/com/toufirm/Country.java
@@ -1,7 +1,6 @@
 package com.toufirm;
 
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
@@ -15,7 +14,7 @@ import lombok.ToString;
 @ToString
 public class Country {
 
-    private Long id;
+    private Integer id;
 
     private String countryName;
 
@@ -23,7 +22,7 @@ public class Country {
 
     }
 
-    public Country(Long id, String countryName) {
+    public Country(Integer id, String countryName) {
         this.id = id;
         this.countryName = countryName;
     }
